+++ react/__mocks__/vtex.product-summary.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react'
+
+export class ProductSummary extends Component {
+  render() {
+    return this.props.children
+  }
+}