+++ react/__mocks__/vtex.styleguide.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+export class Dropdown extends Component {
+  render() {
+    return (<div>Dropdown</div>)
+  }
+}
+
+export class Spinner extends Component {
+  render() {
+    return (<div>Spinner</div>)
+  }
+}