+++ react/constants/CSSClasses.js
@@ -1,0 +1,11 @@
+export default {
+  MAIN_CLASS: 'vtex-search-result',
+  ITEM_CLASS: 'vtex-search-result__item',
+  HEADER_CLASS: 'vtex-search-result__header',
+  HEADER_SEARCH_CLASS: 'vtex-search-result__header-search',
+  HEADER_ORDER_RESULTS_CLASS: 'vtex-search-result__header-order-results',
+  SEARCH_FILTER_MAIN_CLASS: 'vtex-search-result__filter',
+  SEARCH_FILTER_HEADER_ICON: 'vtex-search-result__filter-icon',
+  GALLERY_CLASS: 'vtex-gallery',
+  GALLERY_ITEM_CLASS: 'vtex-gallery__item',
+}