+++ react/test-setup.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+window.matchMedia = window.matchMedia || function() {
+  return {
+    matches: false,
+    addListener: function() {},
+    removeListener: function() {},
+  }
+}