+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Internationalization to the schema.
+- `SearchResult` component
+- `SearchFilter` compoenent
+- `SelectedFilters` component
+- Graphql queries of `search` and `facets`
 
 ## [0.2.0] - 2018-05-11
 ### Added